+++ packages/actor-http-fetch/lib/IFetchInitPreprocessor.ts
@@ -3,4 +3,5 @@
  */
 export interface IFetchInitPreprocessor {
   handle: (init: RequestInit) => Promise<RequestInit>;
+  createAbortController: () => Promise<AbortController>;
 }