+++ packages/actor-init-query/lib/ActorInitQueryBase.ts
@@ -111,11 +111,14 @@ export interface IActorInitQueryBaseArgs extends IActorInitArgs {
    *   "lenient": "@comunica/actor-init-query:lenient",
    *   "httpIncludeCredentials": "@comunica/bus-http:include-credentials",
    *   "httpAuth": "@comunica/bus-http:auth",
+   *   "httpTimeout": "@comunica/bus-http:http-timeout",
+   *   "httpBodyTimeout": "@comunica/bus-http:http-body-timeout",
    *   "fetch": "@comunica/bus-http:fetch",
    *   "readOnly": "@comunica/bus-query-operation:readOnly",
    *   "extensionFunctions": "@comunica/actor-init-query:extensionFunctions",
    *   "extensionFunctionCreator": "@comunica/actor-init-query:extensionFunctionCreator",
-   *   "explain": "@comunica/actor-init-query:explain"
+   *   "explain": "@comunica/actor-init-query:explain",
+   *   "unionDefaultGraph": "@comunica/bus-query-operation:unionDefaultGraph"
    * }}
    */
   contextKeyShortcuts: Record<string, string>;