+++ packages/actor-init-query/test/HttpServiceSparqlEndpoint-test.ts
@@ -471,6 +471,117 @@ describe('HttpServiceSparqlEndpoint', () => {
         expect(http.createServer).toBeCalledTimes(1);
         expect(http.createServer).toHaveBeenLastCalledWith((<any> instance).handleRequest.bind());
       });
+
+      it('should end an open connection on shutdown messages', async() => {
+        let shutdownListener: any;
+        (<any> process).on = (event: any, listener: any): void => {
+          if (event === 'message') {
+            shutdownListener = listener;
+          }
+        };
+
+        // Start server
+        await instance.run(stdout, stderr);
+        const server = http.createServer.mock.results[0].value;
+
+        // Open new connection
+        const response1 = new EventEmitter();
+        (<any> response1).end = jest.fn((message, resolve) => resolve());
+        server.emit('request', undefined, response1);
+
+        // Send shutdown message
+        shutdownListener('shutdown');
+
+        expect((<any> response1).end).toHaveBeenCalledWith('!TIMEDOUT!', expect.anything());
+        await new Promise(setImmediate);
+        expect(process.exit).toHaveBeenCalledTimes(1);
+      });
+
+      it('should end multiple open connections on shutdown messages', async() => {
+        let shutdownListener: any;
+        (<any> process).on = (event: any, listener: any): void => {
+          if (event === 'message') {
+            shutdownListener = listener;
+          }
+        };
+
+        // Start server
+        await instance.run(stdout, stderr);
+        const server = http.createServer.mock.results[0].value;
+
+        // Open new connections
+        const response1 = new EventEmitter();
+        (<any> response1).end = jest.fn((message, resolve) => resolve());
+        server.emit('request', undefined, response1);
+        const response2 = new EventEmitter();
+        (<any> response2).end = jest.fn((message, resolve) => resolve());
+        server.emit('request', undefined, response2);
+
+        // Send shutdown message
+        shutdownListener('shutdown');
+
+        await new Promise(setImmediate);
+        expect((<any> response1).end).toHaveBeenCalledTimes(1);
+        expect((<any> response1).end).toHaveBeenCalledWith('!TIMEDOUT!', expect.anything());
+        expect((<any> response2).end).toHaveBeenCalledTimes(1);
+        expect((<any> response2).end).toHaveBeenCalledWith('!TIMEDOUT!', expect.anything());
+        expect(process.exit).toHaveBeenCalledTimes(1);
+      });
+
+      it('should end only actively open connections on shutdown messages', async() => {
+        let shutdownListener: any;
+        (<any> process).on = (event: any, listener: any): void => {
+          if (event === 'message') {
+            shutdownListener = listener;
+          }
+        };
+
+        // Start server
+        await instance.run(stdout, stderr);
+        const server = http.createServer.mock.results[0].value;
+
+        // Open new connections
+        const response1 = new EventEmitter();
+        (<any> response1).end = jest.fn((message, resolve) => resolve());
+        server.emit('request', undefined, response1);
+        const response2 = new EventEmitter();
+        (<any> response2).end = jest.fn((message, resolve) => resolve());
+        server.emit('request', undefined, response2);
+        response2.emit('close');
+
+        // Send shutdown message
+        shutdownListener('shutdown');
+
+        expect((<any> response1).end).toHaveBeenCalledWith('!TIMEDOUT!', expect.anything());
+        expect((<any> response2).end).not.toHaveBeenCalled();
+        await new Promise(setImmediate);
+        expect(process.exit).toHaveBeenCalledTimes(1);
+      });
+
+      it('should not end open connections on non-shutdown messages', async() => {
+        let shutdownListener: any;
+        (<any> process).on = (event: any, listener: any): void => {
+          if (event === 'message') {
+            shutdownListener = listener;
+          }
+        };
+
+        // Start server
+        await instance.run(stdout, stderr);
+        const server = http.createServer.mock.results[0].value;
+
+        // Open new connection
+        const response1 = new EventEmitter();
+        (<any> response1).end = jest.fn((message, resolve) => resolve());
+        server.emit('request', undefined, response1);
+
+        // Send shutdown message
+        shutdownListener('non-shutdown');
+
+        expect((<any> response1).end).not.toHaveBeenCalled();
+        await new Promise(setImmediate);
+        expect(process.exit).not.toHaveBeenCalled();
+      });
     });
 
     describe('run as a master', () => {
@@ -526,13 +637,14 @@ describe('HttpServiceSparqlEndpoint', () => {
         // Simulate listening event
         const dummyWorker: any = new EventEmitter();
         dummyWorker.send = jest.fn();
+        dummyWorker.isConnected = jest.fn(() => true);
         dummyWorker.process = {
           pid: 123,
         };
         (<any> jest.mocked(cluster.on).mock.calls[0][1])(dummyWorker);
 
         // Simulate start event
-        dummyWorker.emit('message', 'start');
+        dummyWorker.emit('message', { type: 'start', queryId: 0 });
 
         expect(setTimeout).toHaveBeenCalledTimes(1);
         expect(setTimeout).toHaveBeenLastCalledWith(expect.any(Function), 60_000);
@@ -550,20 +662,21 @@ describe('HttpServiceSparqlEndpoint', () => {
         // Simulate listening event
         const dummyWorker: any = new EventEmitter();
         dummyWorker.send = jest.fn();
+        dummyWorker.isConnected = jest.fn(() => true);
         dummyWorker.process = {
           pid: 123,
         };
         (<any> jest.mocked(cluster.on).mock.calls[0][1])(dummyWorker);
 
         // Simulate start event
-        dummyWorker.emit('message', 'start');
+        dummyWorker.emit('message', { type: 'start', queryId: 0 });
 
         expect(setTimeout).toHaveBeenCalledTimes(1);
         expect(setTimeout).toHaveBeenLastCalledWith(expect.any(Function), 60_000);
         expect(dummyWorker.send).not.toHaveBeenCalled();
 
         // Simulate end event
-        dummyWorker.emit('message', 'end');
+        dummyWorker.emit('message', { type: 'end', queryId: 0 });
 
         expect(clearTimeout).toHaveBeenCalledTimes(1);
 
@@ -581,13 +694,14 @@ describe('HttpServiceSparqlEndpoint', () => {
         // Simulate listening event
         const dummyWorker: any = new EventEmitter();
         dummyWorker.send = jest.fn();
+        dummyWorker.isConnected = jest.fn(() => true);
         dummyWorker.process = {
           pid: 123,
         };
         (<any> jest.mocked(cluster.on).mock.calls[0][1])(dummyWorker);
 
         // Simulate start event
-        dummyWorker.emit('message', 'start');
+        dummyWorker.emit('message', { type: 'start', queryId: 0 });
 
         expect(setTimeout).toHaveBeenCalledTimes(1);
         expect(setTimeout).toHaveBeenLastCalledWith(expect.any(Function), 60_000);
@@ -611,6 +725,58 @@ describe('HttpServiceSparqlEndpoint', () => {
         (<any> jest.mocked(process.once).mock.calls[0][1])();
 
         expect(cluster.disconnect).toBeCalledTimes(1);
+      });
+
+      it('should catch worker shutdown errors', async() => {
+        await instance.run(stdout, stderr);
+
+        // Simulate listening event
+        const dummyWorker: any = new EventEmitter();
+        dummyWorker.send = jest.fn(() => {
+          throw new Error('shutdown exception');
+        });
+        dummyWorker.isConnected = jest.fn(() => true);
+        dummyWorker.process = {
+          pid: 123,
+        };
+        (<any> jest.mocked(cluster.on).mock.calls[0][1])(dummyWorker);
+
+        // Simulate start event
+        dummyWorker.emit('message', { type: 'start', queryId: 0 });
+
+        expect(setTimeout).toHaveBeenCalledTimes(1);
+        expect(setTimeout).toHaveBeenLastCalledWith(expect.any(Function), 60_000);
+        expect(dummyWorker.send).not.toHaveBeenCalled();
+
+        // Simulate timeout is passed
+        jest.runAllTimers();
+
+        expect(dummyWorker.send).toHaveBeenCalledWith('shutdown');
+      });
+
+      it('should not send message to disconnected worker', async() => {
+        await instance.run(stdout, stderr);
+
+        // Simulate listening event
+        const dummyWorker: any = new EventEmitter();
+        dummyWorker.send = jest.fn();
+        dummyWorker.isConnected = jest.fn(() => false);
+        dummyWorker.process = {
+          pid: 123,
+        };
+        (<any> jest.mocked(cluster.on).mock.calls[0][1])(dummyWorker);
+
+        // Simulate start event
+        dummyWorker.emit('message', { type: 'start', queryId: 0 });
+
+        expect(setTimeout).toHaveBeenCalledTimes(1);
+        expect(setTimeout).toHaveBeenLastCalledWith(expect.any(Function), 60_000);
+        expect(dummyWorker.send).not.toHaveBeenCalled();
+
+        // Simulate timeout is passed
+        jest.runAllTimers();
+
+        expect(dummyWorker.send).not.toHaveBeenCalled();
       });
     });
 
@@ -643,7 +809,7 @@ describe('HttpServiceSparqlEndpoint', () => {
           await instance.handleRequest(engine, variants, stdout, stderr, request, response);
 
           expect(instance.writeQueryResult)
-            .toHaveBeenCalledWith(engine, stdout, stderr, request, response, undefined, null, false, true);
+            .toHaveBeenCalledWith(engine, stdout, stderr, request, response, undefined, null, false, true, 0);
         });
 
       it('should use the parsed query string when the request method equals GET'
@@ -659,7 +825,8 @@ describe('HttpServiceSparqlEndpoint', () => {
             { type: 'query', value: 'test_query' },
             null,
             false,
-            true);
+            true,
+            0);
         });
 
       it('should set headonly and use the empty query string when the request method is HEAD and url parsing fails'
@@ -676,7 +843,8 @@ describe('HttpServiceSparqlEndpoint', () => {
             undefined,
             null,
             true,
-            true);
+            true,
+            0);
         });
 
       it('should set headonly and use the parsed query string when the request method is HEAD'
@@ -692,7 +860,8 @@ describe('HttpServiceSparqlEndpoint', () => {
             { type: 'query', value: 'test_query' },
             null,
             true,
-            true);
+            true,
+            0);
         });
 
       it('should call writeQueryResult with correct arguments if request method equals POST', async() => {
@@ -708,7 +877,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           { type: 'query', value: 'test_parseBody_result' },
           null,
           false,
-          false);
+          false,
+          0);
       });
 
       it('should choose a mediaType if accept header is set', async() => {
@@ -728,7 +898,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           { type: 'query', value: 'test_parseBody_result' },
           chosen,
           false,
-          false);
+          false,
+          0);
       });
 
       it('should choose the best matching mediaType when we can exactly match', async() => {
@@ -747,7 +918,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           { type: 'query', value: 'test_parseBody_result' },
           'a/a',
           false,
-          false);
+          false,
+          0);
       });
 
       it('should choose the best matching mediaType when we can exactly match with out-of-order q', async() => {
@@ -766,7 +938,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           { type: 'query', value: 'test_parseBody_result' },
           'a/a',
           false,
-          false);
+          false,
+          0);
       });
 
       it('should choose the second best matching mediaType when we can exactly match', async() => {
@@ -785,7 +958,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           { type: 'query', value: 'test_parseBody_result' },
           'b/b',
           false,
-          false);
+          false,
+          0);
       });
 
       it('should choose the mediaType when the first is unknown', async() => {
@@ -804,7 +978,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           { type: 'query', value: 'test_parseBody_result' },
           'b/b',
           false,
-          false);
+          false,
+          0);
       });
 
       it('should choose a null media type if accept header is *', async() => {
@@ -822,7 +997,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           { type: 'query', value: 'test_parseBody_result' },
           null,
           false,
-          false);
+          false,
+          0);
       });
 
       it('should choose a null media type if accept header is */*', async() => {
@@ -840,7 +1016,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           { type: 'query', value: 'test_parseBody_result' },
           null,
           false,
-          false);
+          false,
+          0);
       });
 
       it('should choose a null mediaType if accept header is not set', async() => {
@@ -858,7 +1035,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           { type: 'query', value: 'test_parseBody_result' },
           null,
           false,
-          false);
+          false,
+          0);
       });
 
       it('should only invalidate cache if invalidateCacheBeforeQuery is set to true', async() => {
@@ -938,7 +1116,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           query,
           mediaType,
           false,
-          true);
+          true,
+          0);
 
         await expect(endCalledPromise).resolves.toBe('Rejected query');
         expect(response.writeHead).toHaveBeenLastCalledWith(400,
@@ -956,7 +1135,8 @@ describe('HttpServiceSparqlEndpoint', () => {
             query,
             mediaType,
             false,
-            true);
+            true,
+            0);
 
           await expect(endCalledPromise).resolves.toBe(
             'The response for the given query could not be serialized for the requested media type\n',
@@ -974,7 +1154,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           query,
           mediaType,
           false,
-          true);
+          true,
+          0);
 
         await expect(endCalledPromise).resolves.toBeFalsy();
         expect(response.writeHead).toHaveBeenCalledTimes(1);
@@ -994,7 +1175,8 @@ describe('HttpServiceSparqlEndpoint', () => {
             query,
             mediaType,
             false,
-            true);
+            true,
+            0);
 
           await expect(endCalledPromise).resolves.toBe('An internal server error occurred.\n');
           expect(response.writeHead).toHaveBeenCalledTimes(1);
@@ -1011,7 +1193,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           query,
           mediaType,
           true,
-          true);
+          true,
+          0);
 
         expect(response.writeHead).toHaveBeenCalledTimes(1);
         expect(response.writeHead).toHaveBeenLastCalledWith(200,
@@ -1036,7 +1219,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           { type: 'query', value: '' },
           mediaType,
           false,
-          true);
+          true,
+          0);
 
         // Check output
         await expect(endCalledPromise).resolves.toBeFalsy();
@@ -1087,7 +1271,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           { type: 'query', value: '' },
           mediaType,
           true,
-          true);
+          true,
+          0);
 
         // Check output
         await expect(endCalledPromise).resolves.toBeFalsy();
@@ -1114,7 +1299,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           { type: 'query', value: '' },
           mediaType,
           false,
-          true);
+          true,
+          0);
 
         await expect(endCalledPromise).resolves.toBe('An internal server error occurred.\n');
         expect(response.writeHead).toHaveBeenCalledTimes(1);
@@ -1132,7 +1318,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           { type: 'query', value: '' },
           'mediatype_throwerror',
           false,
-          true);
+          true,
+          0);
 
         await expect(endCalledPromise).resolves.toBe(
           'The response for the given query could not be serialized for the requested media type\n',
@@ -1153,7 +1340,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           query,
           '',
           false,
-          true);
+          true,
+          0);
 
         await expect(endCalledPromise).resolves.toBeFalsy();
         expect(response.writeHead).toHaveBeenCalledTimes(1);
@@ -1174,7 +1362,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           query,
           '',
           false,
-          true);
+          true,
+          0);
 
         await expect(endCalledPromise).resolves.toBeFalsy();
         expect(response.writeHead).toHaveBeenCalledTimes(1);
@@ -1195,7 +1384,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           query,
           '',
           false,
-          true);
+          true,
+          0);
 
         await expect(endCalledPromise).resolves.toBeFalsy();
         expect(response.writeHead).toHaveBeenCalledTimes(1);
@@ -1217,69 +1407,15 @@ describe('HttpServiceSparqlEndpoint', () => {
           query,
           '',
           false,
-          true);
+          true,
+          0);
 
         expect(process.send).toHaveBeenCalledTimes(1);
-        expect(process.send).toHaveBeenCalledWith('start');
+        expect(process.send).toHaveBeenCalledWith({ type: 'start', queryId: 0 });
 
         response.emit('close');
         expect(process.send).toHaveBeenCalledTimes(2);
-        expect(process.send).toHaveBeenCalledWith('end');
-      });
-
-      it('should handle shutdown messages', async() => {
-        let shutdownListener: any;
-        (<any> process).on = (event: any, listener: any): void => {
-          if (event === 'message') {
-            shutdownListener = listener;
-          }
-        };
-        const engine = await new QueryEngineFactoryBase().create();
-        engine.query = () => ({ resultType: 'bindings' });
-
-        await instance.writeQueryResult(engine,
-          new PassThrough(),
-          new PassThrough(),
-          request,
-          response,
-          query,
-          '',
-          false,
-          true);
-
-        expect(process.exit).not.toHaveBeenCalled();
-
-        shutdownListener('shutdown');
-
-        expect(process.exit).toHaveBeenCalledTimes(1);
-        expect(response.end).toHaveBeenCalledWith('!TIMED OUT!');
-      });
-
-      it('should ignore non-shutdown messages', async() => {
-        let shutdownListener: any;
-        (<any> process).on = (event: any, listener: any): void => {
-          if (event === 'message') {
-            shutdownListener = listener;
-          }
-        };
-        const engine = await new QueryEngineFactoryBase().create();
-        engine.query = () => ({ resultType: 'bindings' });
-
-        await instance.writeQueryResult(engine,
-          new PassThrough(),
-          new PassThrough(),
-          request,
-          response,
-          query,
-          '',
-          false,
-          true);
-
-        expect(process.exit).not.toHaveBeenCalled();
-
-        shutdownListener('other');
-
-        expect(process.exit).not.toHaveBeenCalled();
+        expect(process.send).toHaveBeenCalledWith({ type: 'end', queryId: 0 });
       });
 
       it('should fallback to simple for updates if media type is falsy', async() => {
@@ -1294,7 +1430,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           query,
           '',
           false,
-          true);
+          true,
+          0);
 
         await expect(endCalledPromise).resolves.toBeFalsy();
         expect(response.writeHead).toHaveBeenCalledTimes(1);
@@ -1315,7 +1452,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           query,
           '',
           false,
-          true);
+          true,
+          0);
 
         await expect(endCalledPromise).resolves.toBeFalsy();
         expect(engine.query).toHaveBeenCalledWith('default_test_query', { [KeysQueryOperation.readOnly.name]: true });
@@ -1333,7 +1471,8 @@ describe('HttpServiceSparqlEndpoint', () => {
           query,
           '',
           false,
-          false);
+          false,
+          0);
 
         await expect(endCalledPromise).resolves.toBeFalsy();
         expect(engine.query).toHaveBeenCalledWith('default_test_query', {});
@@ -1353,25 +1492,43 @@ describe('HttpServiceSparqlEndpoint', () => {
       });
 
       it('should not error when eventEmitter is undefined', async() => {
-        instance.stopResponse(response);
+        instance.stopResponse(response, 0);
         response.emit('close');
       });
 
       it('should do nothing when no timeout or close event occurs', async() => {
-        instance.stopResponse(response, eventEmitter);
+        instance.stopResponse(response, 0, eventEmitter);
         // Not waiting for timeout to occur
 
         expect(eventEmitter.listeners('test').length).toEqual(1);
         expect(response.end).not.toHaveBeenCalled();
         expect(endListener).not.toHaveBeenCalled();
+        expect(process.exit).not.toHaveBeenCalled();
       });
 
       it('should remove event eventlisteners from eventEmitter when response is closed', async() => {
-        instance.stopResponse(response, eventEmitter);
+        instance.stopResponse(response, 0, eventEmitter);
         response.emit('close');
 
         expect(eventEmitter.listeners('test').length).toEqual(0);
         expect(response.end).toHaveBeenCalled();
+        expect(process.exit).not.toHaveBeenCalled();
+      });
+
+      it('should void error events', async() => {
+        instance.stopResponse(response, 0, eventEmitter);
+        response.emit('close');
+
+        eventEmitter.emit('error', new Error('Ignored stopResponse error'));
+      });
+
+      it('should exit when freshWorkerPerQuery is enabled', async() => {
+        instance = new HttpServiceSparqlEndpoint({ ...argsDefault, workers: 4, freshWorkerPerQuery: true });
+
+        instance.stopResponse(response, 0, eventEmitter);
+        response.emit('close');
+
+        expect(process.exit).toHaveBeenCalledWith(15);
       });
     });