+++ packages/actor-query-operation-quadpattern/lib/ActorQueryOperationQuadpattern.ts
@@ -2,6 +2,7 @@ import { BindingsFactory } from '@comunica/bindings-factory';
 import type { IActionQueryOperation } from '@comunica/bus-query-operation';
 import { ActorQueryOperationTyped } from '@comunica/bus-query-operation';
 import type { MediatorRdfResolveQuadPattern } from '@comunica/bus-rdf-resolve-quad-pattern';
+import { KeysQueryOperation } from '@comunica/context-entries';
 import type { IActorArgs, IActorTest } from '@comunica/core';
 import type { BindingsStream,
   IQueryOperationResult,
@@ -14,9 +15,11 @@ import { DataFactory } from 'rdf-data-factory';
 import type { QuadTermName } from 'rdf-terms';
 import { getTerms, QUAD_TERM_NAMES, reduceTerms, TRIPLE_TERM_NAMES, uniqTerms } from 'rdf-terms';
 import type { Algebra } from 'sparqlalgebrajs';
+import { Factory } from 'sparqlalgebrajs';
 
 const BF = new BindingsFactory();
 const DF = new DataFactory();
+const AF = new Factory();
 
 /**
  * A comunica actor for handling 'quadpattern' query operations.
@@ -24,6 +27,7 @@ const DF = new DataFactory();
 export class ActorQueryOperationQuadpattern extends ActorQueryOperationTyped<Algebra.Pattern>
   implements IActorQueryOperationQuadpatternArgs {
   public readonly mediatorResolveQuadPattern: MediatorRdfResolveQuadPattern;
+  public readonly unionDefaultGraph: boolean;
 
   public constructor(args: IActorQueryOperationQuadpatternArgs) {
     super(args, 'pattern');
@@ -193,8 +197,20 @@ export class ActorQueryOperationQuadpattern extends ActorQueryOperationTyped<Alg
       context = context.merge(pattern.context);
     }
 
+    // Modify pattern with default graph when using union default graph semantics
+    let patternInner = pattern;
+    const unionDefaultGraph = this.unionDefaultGraph || context.get(KeysQueryOperation.unionDefaultGraph);
+    if (pattern.graph.termType === 'DefaultGraph' && unionDefaultGraph) {
+      patternInner = AF.createPattern(
+        pattern.subject,
+        pattern.predicate,
+        pattern.object,
+        DF.variable('__comunica:defaultGraph'),
+      );
+    }
+
     // Resolve the quad pattern
-    const result = await this.mediatorResolveQuadPattern.mediate({ pattern, context });
+    const result = await this.mediatorResolveQuadPattern.mediate({ pattern: patternInner, context });
 
     // Collect all variables from the pattern
     const variables = ActorQueryOperationQuadpattern.getVariables(pattern);
@@ -228,6 +244,12 @@ export class ActorQueryOperationQuadpattern extends ActorQueryOperationTyped<Alg
       const duplicateElementLinks: Record<string, string[]> | undefined = ActorQueryOperationQuadpattern
         .getDuplicateElementLinks(pattern);
 
+      // SPARQL query semantics allow graph variables to only match with named graphs, excluding the default graph
+      // But this is not the case when using union default graph semantics
+      if (pattern.graph.termType === 'Variable' && !unionDefaultGraph) {
+        filteredOutput = filteredOutput.filter(quad => quad.graph.termType !== 'DefaultGraph');
+      }
+
       // If there are duplicate variables in the search pattern,
       // make sure that we filter out the triples that don't have equal values for those triple elements,
       // as QPF ignores variable names.
@@ -258,4 +280,10 @@ export interface IActorQueryOperationQuadpatternArgs extends
    * The quad pattern resolve mediator
    */
   mediatorResolveQuadPattern: MediatorRdfResolveQuadPattern;
+  /**
+   * If the default graph should also contain the union of all named graphs.
+   * This can be overridden by {@link KeysQueryOperation#unionDefaultGraph}.
+   * @default {false}
+   */
+  unionDefaultGraph: boolean;
 }